+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - [#39](https://github.com/rafaelldi/aspire-plugin/issues/39): Trace diagram can be broken because of the deprecated span attributes
 - Fixed show Aspire dashboard action.
+- [#24](https://github.com/rafaelldi/aspire-plugin/issues/24): Substitute environment variables for all Aspire projects 
 
 ### Changed